requirements.txt
streamlit
python-dotenv
google-generativeai
pandas
numpy
plotly
sentence-transformers
scikit-learn
bcrypt
pillow